Update rongyi commit

+++ apps/explorer/lib/explorer/chain/address.ex
@@ -93,7 +93,7 @@ defmodule Explorer.Chain.Address do
     field(:fetched_coin_balance, Wei)
     field(:fetched_coin_balance_block_number, :integer)
     field(:contract_code, Data)
-    field(:nonce, :integer)
+    field(:nonce, :decimal)
     field(:decompiled, :boolean, default: false)
     field(:verified, :boolean, default: false)
     field(:has_decompiled_code?, :boolean, virtual: true)
